created base class from Page and moved numbered page functionality to new NumberedPage class
