+++ trustar/models/enclave.py
@@ -1,8 +1,6 @@
 # python 2 backwards compatibility
 from __future__ import print_function
 from builtins import object, super
-from future import standard_library
-from six import string_types
 
 # package imports
 from .base import ModelBase
